requirements.txt
# Core Testing
selenium==4.15.0
pytest==7.4.0
webdriver-manager==4.0.0
pytest-xdist==3.3.1

# Advanced Reporting (latest compatible versions)
pytest-html==4.1.1
allure-pytest==2.13.2
allure-python-commons==2.13.2